Warnings Removal: removing a warning/error only found by linux.clang on test server
----
An additional 14 commit messages were squashed into this commit:
--
Warnings Removal: fixing a very large number of new warnings/errors

Please be sure to compile in master after a merge of your branch to avoid this
sort of problem.
--
Warnings Removal: another Rhiju warning/error
--
Warnings Removal: another 2 Rhiju warnings/errors
--
Warnings Removal: another 6 Rhiju warnings/errors
--
Warnings Removal: 3 more Rhiju warnings/errors
--
Warnings Removal: 2 more Rhiju warnings/errors
--
Warnings Removal: 5 more Rhiju warnings/errors
--
Warnings Removal: 5 more errors from Rhiju's code
--
Warnings Removal: 4 more errors from Rhiju's code
--
Warnings Removal: 2 more errors from Rhiju's code
--
Warnings Removal: mistake in fixing warnings
--
Warnings Removal: fixing warnings introduced by Rhiju and updating cmake settings
--
Warnings Removal: fixing warnings that should have been wrapped with an ifdef
--
Warnings Removal: updating cmake settings
--

Old repository SHA1: 342a8bbee78475c476e52031408dd894777cd68e

+++ source/src/protocols/rotamer_sampler/rigid_body/FloatingBaseUtil.cc
@@ -9,7 +9,7 @@
 
 /// @file protocols/rotamer_sampler/rigid_body/FloatingBaseUtil.cc
 /// @brief
-/// @detailed
+/// @details
 /// @author Rhiju Das, rhiju@stanford.edu
 
 
@@ -86,7 +86,7 @@ namespace rigid_body {
 																		 Size const moving_res ){
 		kinematics::FoldTree const & fold_tree = pose.fold_tree();
 		for ( Size n = 1; n <= fold_tree.num_jump(); n++ ) {
-			if ( fold_tree.downstream_jump_residue( n ) == moving_res  ) {
+			if ( fold_tree.downstream_jump_residue( n ) == static_cast< int >( moving_res ) ) {
 				Size const reference_res = fold_tree.upstream_jump_residue( n );
 				return reference_res;
 			}
@@ -102,8 +102,8 @@ namespace rigid_body {
 		Size jump_number( 0 );
 		kinematics::FoldTree const & fold_tree = pose.fold_tree();
 		for ( Size n = 1; n <= fold_tree.num_jump(); n++ ) {
-			if ( fold_tree.downstream_jump_residue( n ) == moving_res &&
-					 fold_tree.upstream_jump_residue( n ) == reference_res ) {
+			if ( fold_tree.downstream_jump_residue( n ) == static_cast< int >( moving_res ) &&
+					fold_tree.upstream_jump_residue( n ) == static_cast< int >( reference_res ) ) {
 				jump_number = n;
 				break;
 			}
